✨ : Move buy-required action into Process component
